+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+
+## v 3.5
+ * Add asq-js-function-body question type
+ 
+## v 3.2
+ * Add asq-css-select question type
+
 ## v 0.1
  * Make seperate repo from ASQ
  * Impress presentation example and asq-previewer is now in this repo